requirements.txt
flask==2.1.1
pandas==1.3.5
boto3==1.21.45
opencv-python==4.5.5.64
pillow==8.0.1